Company Generator/requirements.txt
beautifulsoup4==4.9.1
requests==2.24.0
psycopg2-binary==2.8.5